chore: update dependencies to latest versions
Update major dependencies to resolve future-incompat warnings:

- sqlx: 0.7 → 0.8 (all crates)
- redis: 0.23 → 0.27 (trading-common, trading-core)
- config: 0.13 → 0.14 (trading-core)

Breaking changes fixed:
- redis::Commands::expire() now takes i64 instead of usize

This resolves the Rust 2024 edition compatibility warnings for:
- redis: never type fallback issue in script.rs
- sqlx-postgres: similar future-incompat issues

All tests pass. Build completes successfully.

+++ trading-common/src/data/cache.rs
@@ -251,7 +251,7 @@ impl TickDataCache for RedisTickCache {
 
         // Set TTL
         let _: () = conn
-            .expire(&key, self.ttl_seconds as usize)
+            .expire(&key, self.ttl_seconds as i64)
             .map_err(|e| DataError::Cache(format!("Redis EXPIRE failed: {}", e)))?;
 
         debug!(
